% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{corrcov}
\alias{corrcov}
\title{corrcov}
\usage{
corrcov(z, f, N0, N1, Sigma, thr, W = 0.2, rep = 1000)
}
\arguments{
\item{z}{Marginal z-scores}

\item{f}{Minor allele frequencies}

\item{N0}{Number of controls}

\item{N1}{Number of cases}

\item{Sigma}{SNP correlation matrix}

\item{thr}{Minimum threshold for fine-mapping experiment}

\item{W}{Prior for the standard deviation of the effect size parameter beta}

\item{rep}{The number of simulated posterior probability systems to consider for the corrected coverage estimate. Simulations show that rep=1000 is a robust choice.}
}
\value{
Corrected coverage estimate
}
\description{
Obtain corrected coverage estimate
}
\details{
This function only requires the marginal summary statistics from GWAS
}
