% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{corrcov_bhat}
\alias{corrcov_bhat}
\title{Corrected coverage estimate using estimated effect sizes and their standard errors}
\usage{
corrcov_bhat(bhat, V, N0, N1, Sigma, thr = 0.95, W = 0.2,
  nrep = 1000)
}
\arguments{
\item{bhat}{Estimated effect sizes from single-SNP logistic regressions}

\item{V}{Variance of estimated effect sizes}

\item{N0}{Number of controls}

\item{N1}{Number of cases}

\item{Sigma}{SNP correlation matrix}

\item{thr}{Minimum threshold for fine-mapping experiment (default is 0.95)}

\item{W}{Prior for the standard deviation of the effect size parameter beta}

\item{nrep}{The number of simulated posterior probability systems to consider for the corrected coverage estimate (nrep = 1000 default)}
}
\value{
Corrected coverage estimate
}
\description{
Corrected coverage estimate using estimated effect sizes and their standard errors
}
\details{
This function only requires the marginal summary statistics from GWAS
}
\examples{

set.seed(1)
nsnps <- 100
N0 <- 1000 # number of controls
N1 <- 1000 # number of cases

# simulate fake haplotypes to obtain MAFs and LD matrix
nhaps <- 1000
lag <- 5
maf.tmp <- runif(nsnps+lag, 0.05, 0.5) # common SNPs
laghaps <- do.call("cbind", lapply(maf.tmp, function(f) rbinom(nhaps,1,f)))
haps <- laghaps[,1:nsnps]
for(j in 1:lag)
   haps <- haps + laghaps[,(1:nsnps)+j]
haps <- round(haps/matrix(apply(haps,2,max),nhaps,nsnps,byrow=TRUE))
maf <- colMeans(haps)
LD <- cor2(haps)

varbeta <- Var.data.cc(f = maf, N = N0 + N1, s = N1/(N0+N1))

bhats = rnorm(nsnps, 0, 0.2) # log OR


corrcov_bhat(bhat = bhats, V = varbeta, N0, N1, Sigma = LD)

}
\author{
Anna Hutchinson
}
