% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{corrcov_CI}
\alias{corrcov_CI}
\title{Obtain confidence interval for corrected coverage estimate using Z-scores and mafs}
\usage{
corrcov_CI(z, f, N0, N1, Sigma, thr = 0.95, W = 0.2, nrep = 1000,
  CI = 0.95)
}
\arguments{
\item{z}{Marginal Z-scores}

\item{f}{Minor allele frequencies}

\item{N0}{Number of controls}

\item{N1}{Number of cases}

\item{Sigma}{SNP correlation matrix}

\item{thr}{Minimum threshold for fine-mapping experiment (default is 0.95)}

\item{W}{Prior for the standard deviation of the effect size parameter beta}

\item{nrep}{The number of simulated posterior probability systems to consider for the corrected coverage estimate (nrep = 1000 default)}

\item{CI}{The size of the confidence interval (as a decimal)}
}
\value{
Corrected coverage estimate
}
\description{
Obtain confidence interval for corrected coverage estimate using Z-scores and mafs
}
