% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{corrcov_CI_bhat}
\alias{corrcov_CI_bhat}
\title{Confidence interval for corrected coverage estimate using estimated effect sizes and their standard errors}
\usage{
corrcov_CI_bhat(bhat, V, N0, N1, Sigma, thr = 0.95, W = 0.2,
  nrep = 1000, CI = 0.95)
}
\arguments{
\item{bhat}{Estimated effect sizes from single-SNP logistic regressions}

\item{V}{Variance of estimated effect sizes}

\item{N0}{Number of controls}

\item{N1}{Number of cases}

\item{Sigma}{SNP correlation matrix}

\item{thr}{Minimum threshold for fine-mapping experiment (default is 0.95)}

\item{W}{Prior for the standard deviation of the effect size parameter beta}

\item{nrep}{The number of simulated posterior probability systems to consider for the corrected coverage estimate (nrep = 1000 default)}

\item{CI}{The size of the confidence interval (as a decimal)}
}
\value{
Corrected coverage estimate
}
\description{
Obtain confidence interval for corrected coverage estimate using estimated effect sizes and their standard errors
}
