% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{corrected_cov}
\alias{corrected_cov}
\title{corrected_cov}
\usage{
corrected_cov(mu, nsnps = 200, V, Sigma, pp0, thresh)
}
\arguments{
\item{mu}{The true effect at the CV}

\item{nsnps}{Number of SNPs}

\item{V}{Variance of the estimated effect size (can be obtained using var.beta.cc function)}

\item{Sigma}{SNP correlation matrix}

\item{pp0}{Posterior probabilities of system of interest}

\item{thresh}{Minimum threshold for fine-mapping experiment}
}
\value{
Corrected coverage estimate
}
\description{
Provide a corrected coverage estimate of the causal variant in the credible set
}
\details{
Requires an estimate of the true effect at the CV
}
