% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cs.R
\name{corrected_cs_bhat}
\alias{corrected_cs_bhat}
\title{corrected_cs_bhat}
\usage{
corrected_cs_bhat(bhat, V, N0, N1, lower, upper, desired.cov,
  acc = 5e-04, max.iter = 20)
}
\arguments{
\item{bhat}{Estimated effect sizes}

\item{V}{Prior variance of estimated effect sizes}

\item{N0}{Number of controls}

\item{N1}{Number of cases}

\item{lower}{Lower threshold}

\item{upper}{Upper threshold}

\item{desired.cov}{The desired coverage of the causal variant in the credible set}

\item{acc}{Accuracy of corrected coverage to desired coverage (default = 0.0005)}

\item{max.iter}{Maximum iterations (default = 20)}
}
\value{
list of variants in credible set, required threshold, the corrected coverage and the size of the credible set
}
\description{
corrected_cs_bhat
}
