% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{credset}
\alias{credset}
\title{credset}
\usage{
credset(pp, CV = iCV, thr)
}
\arguments{
\item{pp}{Vector of posterior probabilities}

\item{CV}{Index of CV}

\item{thr}{Minimum threshold for credible set size}
}
\value{
data.frame of claimed.cov (cumulative sum of the posterior probabilities of the variants forming the credible set), binary covered indicator (1 if CV is contained in the credible set), nvar (number of variants in the set)
}
\description{
Obtain a credible set using the Bayesian approach for fine-mapping ([Maller et al., 2012](https://www.ncbi.nlm.nih.gov/pubmed/23104008)).
}
