% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{credset3}
\alias{credset3}
\title{Quicker credset function for matrix of posterior probabilities (using Cpp)}
\usage{
credset3(pp, CV = iCV, thr = 0.6)
}
\arguments{
\item{pp}{Matrix of posterior probabilities (row per system)}

\item{CV}{Vector of indices of the causal variant for each row of posterior probabilities}

\item{thr}{Threshold}
}
\value{
Data.frame of claimed coverage (sum of posterior probabilities of variants in the set), binary covered indicator and number of variants.
}
\description{
Quicker credset function for matrix of posterior probabilities (using Cpp)
}
