% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{credsetC}
\alias{credsetC}
\title{Credible set of variants from matrix of PPs}
\usage{
credsetC(pp, CV, thr = 0.95)
}
\arguments{
\item{pp}{Matrix of posterior probabilities of causality (one row per system)}

\item{CV}{Vector of CV indices (one per system/row)}

\item{thr}{Minimum threshold for credible set size (default is 0.95)}
}
\value{
Data.frame of claimed coverage (sum of posterior probabilities of variants in the set), binary covered indicator and number of variants (nvar).
}
\description{
Quick credset function for matrix of posterior probabilities (using RCpp)
}
