% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{credsetmat}
\alias{credsetmat}
\title{Obtain credible sets from a matrix of posterior probabilities}
\usage{
credsetmat(pp, iCV, threshold)
}
\arguments{
\item{pp}{Matrix of posterior probabilities (one row for each simulation)}

\item{iCV}{A vector of the indices of the CV}

\item{threshold}{The threshold to use to generate the credible set}
}
\description{
Obtain credible sets from a matrix of posterior probabilities
}
