% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_utils.R
\name{.zj_abf}
\alias{.zj_abf}
\title{Internal function: Simulate nrep ABFs from joint Z-score vector}
\usage{
.zj_abf(Zj, int.Sigma, int.nrep, int.ERR, int.r)
}
\arguments{
\item{Zj}{joint z vector}

\item{int.Sigma}{internal sigma}

\item{int.nrep}{internal nrep}

\item{int.ERR}{internal error matrix}

\item{int.r}{internal r}
}
\value{
Matrix of simulated ABFs, one simulation per row
}
\description{
Does not include posterior probabilities for null model
}
