% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_utils.R
\name{.zj_pp}
\alias{.zj_pp}
\title{Simulate posterior probabilities of causality from joint
    Z-score vector}
\usage{
.zj_pp(Zj, int.Sigma, int.nrep, int.ERR, int.r)
}
\arguments{
\item{Zj}{joint z vector}

\item{int.Sigma}{internal sigma}

\item{int.nrep}{internal nrep}

\item{int.ERR}{internal error matrix}

\item{int.r}{internal r}
}
\value{
Matrix of simulated posterior probabilties of causality,
    one simulation per row
}
\description{
Internal function: Simulate nrep posterior probabilities of causality from joint Z-score vector
}
\details{
Does not include posterior probabilities for null model
}
