% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{est_mu_bhat}
\alias{est_mu_bhat}
\title{Estimate the true effect at the causal variant}
\usage{
est_mu_bhat(bhat, V, N0, N1, W = 0.2)
}
\arguments{
\item{bhat}{Vector of estimated effect sizes}

\item{V}{Prior variance for estimated effect sizes}

\item{N0}{Number of controls}

\item{N1}{Number of cases}

\item{W}{Prior for the standard deviation of the effect size parameter beta}
}
\value{
Estimate of the true effect at the causal variant
}
\description{
Estimate the true effect at the causal variant
}
\author{
Anna Hutchinson
}
