% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{geth}
\alias{geth}
\title{geth}
\usage{
geth(nsnp)
}
\arguments{
\item{nsnp}{Number of SNPs to consider for fine-mapping}
}
\value{
Haplotype matrix
}
\description{
Obtain phased haplotype matrix from UK10K data
}
\details{
This function takes 100 SNPs from one end (likely to be in LD) and 100 SNPs from elsewhere, so that the CV will be in LD with some variants and not at all linked to others.
}
\author{
Chris Wallace
}
