% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_utils.R
\name{logsum_matrix}
\alias{logsum_matrix}
\title{logsum rows of a matrix}
\usage{
logsum_matrix(x)
}
\arguments{
\item{x}{numeric matrix}
}
\value{
rowwise sums
}
\description{
logsum rows of a matrix
}
\author{
Chris Wallace
}
