% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_simulation_functions.R
\name{pp_ERR}
\alias{pp_ERR}
\title{Simulate posterior probabilities of causality from joint Z-score vector}
\usage{
pp_ERR(Zj)
}
\arguments{
\item{Zj}{Vector of joint Z-scores (0s except at CV)}
}
\value{
Matrix of simulated posterior probabilties of causality, one simulation per row
}
\description{
Internal function: Simulate nrep posterior probabilities of causality from joint Z-score vector
}
\details{
Does not include posterior probabilities for null model
}
