% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{ppfunc}
\alias{ppfunc}
\title{Posterior probabilities for vectors}
\usage{
ppfunc(z, V, W = 0.2)
}
\arguments{
\item{z}{Vector of marginal z-scores}

\item{V}{Variance of the estimated effect size (can be obtained using var.beta.cc function)}

\item{W}{Prior for the standard deviation of the effect size parameter $\beta$}
}
\value{
Vector of posterior probabilities
}
\description{
Quick posterior probabilities for each SNP causal from marginal z-scores
}
\details{
This derivation does not include the null model, so one posterior probability is obtained for each marginal
z-score and these sum to 1.
}
