% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{pred_logit}
\alias{pred_logit}
\title{pred_logit}
\usage{
pred_logit(x)
}
\arguments{
\item{x}{data.frame with column for 'covered' and 'logit.claim'}
}
\value{
Predicted probability of covered
}
\description{
Build and predict corrected coverage using logistic GAM
}
