% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{pred_na}
\alias{pred_na}
\title{pred_na}
\usage{
pred_na(x)
}
\arguments{
\item{x}{data.frame with a binary 'covered' column}
}
\value{
Predicted coverage
}
\description{
Prediction for corrected coverage if gam cannot be fitted
}
\author{
Anna Hutchinson
}
