% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{prop_cov}
\alias{prop_cov}
\title{prop_cov}
\usage{
prop_cov(x)
}
\arguments{
\item{x}{data.frame with a binary 'covered' column}
}
\value{
Proportion of x with x = 1
}
\description{
Proportion of simulated credible sets containing the causal variant
}
\author{
Anna Hutchinson
}
