% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{quick_corrcov}
\alias{quick_corrcov}
\title{Use simulated pps to find corrected coverage estimate}
\usage{
quick_corrcov(thr, simulated.pps, pp)
}
\arguments{
\item{thr}{Threshold value to exceed}

\item{simulated.pps}{A list of matrices of simulated posterior probabilities (1 simulation per row) where each list element is for the corresponding SNP considered causal}

\item{pp}{The posterior probabilities of the original system}
}
\value{
Corrected coverage estimate
}
\description{
Use simulated pps to find corrected coverage estimate
}
