% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{z0_pp}
\alias{z0_pp}
\title{z0_pp}
\usage{
z0_pp(z, f, type, N, s, W = 0.2)
}
\arguments{
\item{z}{Marginal Z-scores of SNPs}

\item{f}{Minor allele frequencies}

\item{type}{Type of experiment ("quant" or "cc")}

\item{N}{Total sample size}

\item{s}{Proportion of cases (N1/N0+N1), ignored if type=="quant"}

\item{W}{Prior for the standard deviation of the effect size parameter beta}
}
\value{
Posterior probabilities of null model (no genetic effect) and causality for each SNP
}
\description{
Posterior probabilities of causality from marginal Z-scores with prior SD as a parameter
}
\details{
This function converts Z-scores to posterior probabilities of causality, including the null model of no genetic effects
}
\author{
Anna Hutchinson
}
