% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_simulation_functions.R
\name{zj_pp}
\alias{zj_pp}
\title{Simulate marginal from joint Z-score vector}
\usage{
zj_pp(Zj, V, nrep = 5000, W = 0.2, Sigma)
}
\arguments{
\item{Zj}{Vector of joint z-scores (0s except at CV)}

\item{V}{Variance of the estimated effect size (can be obtained using var.beta.cc function)}

\item{nrep}{Number of marginal z-scores to simulate}

\item{W}{Prior for the standard deviation of the effect size parameter, beta}

\item{Sigma}{SNP correlation matrix}
}
\value{
Matrix of simulated posterior probabilties, one simulation per row
}
\description{
Simulate nrep marginal Z-scores from joint Z-scores and convert these to posterior probabilities of causality
}
\details{
Does not include posterior probabilities for null model s.t the output will sum to 1
}
