% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zj_pp_arma}
\alias{zj_pp_arma}
\title{Obtain pp from a matrix of Zj and ERR}
\usage{
zj_pp_arma(Zj, sigma, nrep, ERR, r)
}
\arguments{
\item{Zj}{vector of joint Z scores}

\item{sigma}{SNP correlation matrix}

\item{nrep}{Number of simulations}

\item{ERR}{Error matrix}

\item{r}{Shrinkage values}
}
\value{
pp Matrix of posterior probabilities (one row for each simulation)
}
\description{
Obtain pp from a matrix of Zj and ERR
}
