% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{approx.bf.p}
\alias{approx.bf.p}
\title{approx.bf.p}
\usage{
approx.bf.p(pvals, f, type, N, s, W = 0.2)
}
\arguments{
\item{pvals}{p-values}

\item{f}{Minor allele frequencies}

\item{type}{Type of experiment ("quant" or "cc")}

\item{N}{Total sample size}

\item{s}{Proportion of cases (N1/N0+N1), ignored if type=="quant"}

\item{W}{Prior for the standard deviation of the effect size parameter beta (W=0.2 default)}
}
\value{
data.frame containing lABF and intermediate calculations
}
\description{
Wakefield's log asymptotic Bayes factor (lABF) with prior standard deviation of effect size as a parameter
}
\details{
([Wakefield et al. 2009](https://onlinelibrary.wiley.com/doi/abs/10.1002/gepi.20359)
This function converts p-values to log ABFs, also reporting intermediate calculations
}
