% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{approx.bf.p}
\alias{approx.bf.p}
\title{approx.bf.p}
\usage{
approx.bf.p(p, f, type, N, s, W)
}
\arguments{
\item{p}{p-values}

\item{f}{Minor allele frequencies}

\item{type}{Type of experiment ("quant" or "cc")}

\item{N}{Total sample size}

\item{s}{Proportion of cases (N1/N0+N1), ignored if type=="quant"}

\item{W}{Prior for the standard deviation of the effect size parameter beta}
}
\value{
data.frame containing lABF and intermediate calculations
}
\description{
Wakefield's ABF with prior SD as a parameter
}
\details{
This function converts p-values to ABF
}
\author{
Chris Wallace
}
