% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{cor2}
\alias{cor2}
\title{Correlation matrix of SNPS}
\usage{
cor2(x)
}
\arguments{
\item{x}{Phased haplotype matrix, rows as samples and columns as SNPs}
}
\value{
Correlation matrix
}
\description{
Correlation matrix of SNPS
}
\details{
Quick function to find a correlation matrix
}
\author{
Chris Wallace
}
