% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{corrcov}
\alias{corrcov}
\title{corrcov}
\usage{
corrcov(z0, f, N0, N1, Sigma)
}
\arguments{
\item{z0}{Marginal z-scores}

\item{f}{Minor allele frequencies}

\item{N0}{Number of controls}

\item{N1}{Number of cases}

\item{Sigma}{SNP correlation matrix}
}
\value{
Corrected coverage estimate
}
\description{
Obtain corrected coverage estimate
}
\details{
This function only requires the marginal summary statistics from GWAS
}
