% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{corrcov_bhat}
\alias{corrcov_bhat}
\title{corrcov_bhat}
\usage{
corrcov_bhat(bhat, V, N0, N1, Sigma, thr, W = 0.2, nrep = 1000)
}
\arguments{
\item{bhat}{Estimated effect sizes from single-SNP logistic regressions}

\item{V}{Variance of estimated effect sizes}

\item{N0}{Number of controls}

\item{N1}{Number of cases}

\item{Sigma}{SNP correlation matrix}

\item{thr}{Minimum threshold for fine-mapping experiment}

\item{W}{Prior for the standard deviation of the effect size parameter beta}

\item{nrep}{The number of simulated posterior probability systems to consider for the corrected coverage estimate (nrep = 1000 default)}
}
\value{
Corrected coverage estimate
}
\description{
Obtain corrected coverage estimate using estimated effect sizes (beta hats) and their standard errors
}
\details{
This function only requires the marginal summary statistics from GWAS
}
