% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{corrcov_CI}
\alias{corrcov_CI}
\title{Confidence interval for corrected coverage estimate using Z-scores and mafs}
\usage{
corrcov_CI(z, f, N0, N1, Sigma, thr = 0.95, W = 0.2, nrep = 1000,
  CI = 0.95)
}
\arguments{
\item{z}{Marginal Z-scores}

\item{f}{Minor allele frequencies}

\item{N0}{Number of controls}

\item{N1}{Number of cases}

\item{Sigma}{SNP correlation matrix}

\item{thr}{Minimum threshold for fine-mapping experiment (default is 0.95)}

\item{W}{Prior for the standard deviation of the effect size parameter beta}

\item{nrep}{The number of simulated posterior probability systems to consider for the corrected coverage estimate (nrep = 1000 default)}

\item{CI}{The size of the confidence interval (as a decimal)}
}
\value{
CI for corrected coverage estimate
}
\description{
Obtain confidence interval for corrected coverage estimate using Z-scores and mafs
}
\examples{

\dontrun{

 # this is a long running example
set.seed(1)
nsnps = 100
N0 = 5000
N1 = 5000
z_scores <- rnorm(nsnps, 0, 3) # simulate a vector of Z-scores

# simulate fake haplotypes to obtain MAFs and LD matrix
nhaps <- 1000
lag <- 5
maf.tmp <- runif(nsnps+lag, 0.05, 0.5) # common SNPs
laghaps <- do.call("cbind", lapply(maf.tmp, function(f) rbinom(nhaps,1,f)))
haps <- laghaps[,1:nsnps]
for(j in 1:lag)
   haps <- haps + laghaps[,(1:nsnps)+j]
haps <- round(haps/matrix(apply(haps,2,max),nhaps,nsnps,byrow=TRUE))
maf <- colMeans(haps)
LD <- cor2(haps)

corrcov_CI(z = z_scores, f = maf, N0, N1, Sigma = LD)
}

}
\author{
Anna Hutchinson
}
