% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{corrcov_nvar}
\alias{corrcov_nvar}
\title{Corrected coverage estimate using Z-scores and mafs (fixing nvar)}
\usage{
corrcov_nvar(z, f, N0, N1, Sigma, nvar, thr = 0.95, W = 0.2,
  nrep = 10000)
}
\arguments{
\item{z}{Marginal Z-scores}

\item{f}{Minor allele frequencies}

\item{N0}{Number of controls}

\item{N1}{Number of cases}

\item{Sigma}{SNP correlation matrix}

\item{nvar}{The number of variants that simulated credible sets used for estimation should contain}

\item{thr}{Minimum threshold for fine-mapping experiment (default is 0.95)}

\item{W}{Prior for the standard deviation of the effect size parameter beta}

\item{nrep}{The number of simulated posterior probability systems to consider for the corrected coverage estimate (nrep = 10000 default because we trim out the ones without correct nvar)}
}
\value{
Corrected coverage estimate
}
\description{
Obtain corrected coverage estimate using Z-scores and mafs (limiting simulations used for estimation to those with correct nvar)
}
\details{
This function requires the marginal summary statistics from GWAS and an nvar value. It should only be used when nvar is very low ($<3$) and there is some evidence to suggest that only simulated credible sets with this nvar value should be used to derive the corrected coverage estimate.
}
