% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{corrected_cov}
\alias{corrected_cov}
\title{corrected_cov}
\usage{
corrected_cov(mu, V, W = 0.2, Sigma, pp0, thresh, nrep = 1000)
}
\arguments{
\item{mu}{The true effect at the CV}

\item{V}{Variance of the estimated effect size (can be obtained using Var.beta.cc function)}

\item{W}{Prior for the standard deviation of the effect size parameter beta (W=0.2 default)}

\item{Sigma}{SNP correlation matrix}

\item{pp0}{Posterior probabilities of system of interest}

\item{thresh}{Minimum threshold for fine-mapping experiment}

\item{nrep}{Number of posterior probability systems to simulate for each variant considered causal (nrep = 1000 default)}
}
\value{
Corrected coverage estimate
}
\description{
Obtain a corrected coverage estimate of the causal variant in the credible set
}
\details{
Requires an estimate of the true effect at the CV (e.g. use maximum absolute z-score or output from corrcoverage::mu_est function)
}
