% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{corrected_cov}
\alias{corrected_cov}
\title{corrected_cov}
\usage{
corrected_cov(mu, V, Sigma, pp0, thresh, size)
}
\arguments{
\item{mu}{An estimate for the true effect at the CV}

\item{V}{Variance of the estimated effect size (can be obtained using var.beta.cc function)}

\item{Sigma}{SNP correlation matrix}

\item{pp0}{Posterior probabilities of system of interest}

\item{thresh}{Minimum threshold for fine-mapping experiment}

\item{size}{Optional parameter: The size of the credible set (sum of the posterior probabilities of the variants). If not supplied, then the function can calculate this}
}
\value{
Corrected coverage estimate
}
\description{
Provide a corrected coverage estimate of the causal variant (CV) in the credible set
}
\details{
Requires an estimate of the true effect at the CV (e.g. use maximum absolute z-score or output from mu_est function)
}
