% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cs.R
\name{corrected_cs}
\alias{corrected_cs}
\title{Get new credible set with desired coverage of the CV}
\usage{
corrected_cs(z, f, N0, N1, Sigma, lower, upper, desired.cov, acc = 0.005,
  max.iter = 20)
}
\arguments{
\item{z}{Z-scores}

\item{f}{Minor allele frequencies}

\item{N0}{Number of controls}

\item{N1}{Number of cases}

\item{Sigma}{Correlation matrix of SNPs}

\item{lower}{Lower threshold}

\item{upper}{Upper threshold}

\item{desired.cov}{The desired coverage of the causal variant in the credible set}

\item{acc}{Accuracy of corrected coverage to desired coverage (default = 0.005)}

\item{max.iter}{Maximum iterations (default = 20)}
}
\value{
list of variants in credible set, required threshold, the corrected coverage and the size of the credible set
}
\description{
Get new credible set with desired coverage of the CV
}
