% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{credset}
\alias{credset}
\title{Get credible set of variants}
\usage{
credset(pp, CV, thr = 0.95)
}
\arguments{
\item{pp}{Vector of posterior probabilities of causality}

\item{CV}{Optional parameter: Index of CV}

\item{thr}{Minimum threshold for credible set size (default is 0.95)}
}
\value{
list of the variants in the credible set, the claimed.cov (cumulative sum of the posterior probabilities of the variants forming the credible set), binary covered indicator (1 if CV is contained in the credible set) and nvar (number of variants in the set)
}
\description{
Get credible set of variants
}
\details{
If the CV parameter is supplied (index of causal variant) then the
output includes a binary indicator of whether the CV is contained in the set
}
