% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{credsetC}
\alias{credsetC}
\title{credsetC}
\usage{
credsetC(pp, CV = iCV, thr = 0.6)
}
\arguments{
\item{pp}{Matrix of posterior probabilities of causality (one row per system)}

\item{CV}{Vector of CV indices (one per system/row)}

\item{thr}{Minimum threshold for credible set size}
}
\value{
Data.frame of claimed coverage (sum of posterior probabilities of variants in the set), binary covered indicator and number of variants (nvar).
}
\description{
Quicker credset function for matrix of posterior probabilities (using RCpp)
}
