% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{geth}
\alias{geth}
\title{geth}
\usage{
geth(nsnp)
}
\arguments{
\item{nsnp}{Number of SNPs to consider for fine-mapping}
}
\value{
Haplotype matrix
}
\description{
Obtain phased haplotype matrix from UK10K data
}
\author{
Chris Wallace
}
