% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{ppfunc.mat}
\alias{ppfunc.mat}
\title{ppfunc.mat}
\usage{
ppfunc.mat(zstar, V, W = 0.2)
}
\arguments{
\item{zstar}{Matrix of marginal z-scores, one replicate per row}

\item{V}{Variance of the estimated effect size (can be obtained using var.beta.cc function), one element per column of zstar}

\item{W}{Prior for the standard deviation of the effect size parameter \beta}
}
\value{
Vector of posterior probabilities
}
\description{
Obtain posterior probabilities for each SNP causal from multiple marginal z-scores simulations
}
\details{
This function converts a matrix of z-scores (one row per simulation) to posterior probabilities, not including the null model of no genetic effects, so that the sum of the posterior probabilities for each simulation (each row) is 1.
}
\author{
Chris Wallace
}
