% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{ppfunc.mat}
\alias{ppfunc.mat}
\title{Posterior probabilities for matrices}
\usage{
ppfunc.mat(zstar, V, W = 0.2)
}
\arguments{
\item{zstar}{Matrix of marginal z-scores, one replicate per row}

\item{V}{Variance of the estimated effect size (can be obtained using var.beta.cc function), one element per column of zstar}

\item{W}{Prior for the standard deviation of the effect size parameter $\beta$}
}
\value{
Vector of posterior probabilities
}
\description{
Obtain posterior probabilities for each SNP causal from multiple marginal z-scores simulations
}
\details{
Each row should be a simulation and each column a SNP
}
\author{
Chris Wallace
}
