% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{ppfunc.mat}
\alias{ppfunc.mat}
\title{Find posterior probabilities of causality from matrix of Z-scores}
\usage{
ppfunc.mat(zstar, V, W = 0.2)
}
\arguments{
\item{zstar}{Matrix of marginal z-scores, one replicate per row}

\item{V}{Variance of the estimated effect size, one element per column of zstar}

\item{W}{Prior for the standard deviation of the effect size parameter, beta}
}
\value{
Matrix of posterior probabilities of causality
}
\description{
Posterior probabilities of causality from matrix of marginal Z-scores (1 simulation per row)
}
\details{
This function converts a matrix of Z-scores (one row per simulation) to posterior probabilities of causality,
not including the null model of no genetic effects,
so that the sum of the posterior probabilities for each simulation (each row) is 1.
}
\author{
Chris Wallace
}
