% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{ppfunc}
\alias{ppfunc}
\title{ppfunc}
\usage{
ppfunc(z, V, W = 0.2)
}
\arguments{
\item{z}{Vector of marginal Z-scores}

\item{V}{Variance of the estimated effect size (can be obtained using Var.beta.cc function)}

\item{W}{Prior for the standard deviation of the effect size parameter, beta (W = 0.2 default)}
}
\value{
Vector of posterior probabilities
}
\description{
Posterior probabilities of causality from marginal Z-scores
}
\details{
This function converts Z-scores to posterior probabilities of causality
i.e. not including the null model of no genetic effects,
so that the sum of the posterior probabilities for all variants is 1
}
