% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{ppfunc}
\alias{ppfunc}
\title{ppfunc}
\usage{
ppfunc(z, V, W = 0.2)
}
\arguments{
\item{z}{Vector of marginal z-scores}

\item{V}{Variance of the estimated effect size (can be obtained using var.beta.cc function)}

\item{W}{Prior for the standard deviation of the effect size parameter \beta}
}
\value{
Vector of posterior probabilities
}
\description{
Quick posterior probabilities for each SNP causal from marginal z-scores
}
\details{
This function converts z-scores to posterior probabilities, not including the null model of no genetic effects, so that the sum of the posterior probabilities is 1
}
