% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{pred_logit}
\alias{pred_logit}
\title{pred_logit}
\usage{
pred_logit(x, size)
}
\arguments{
\item{x}{data.frame with column for 'covered' and 'logit.claim'}

\item{size}{size of the credible set to correct (sum of the posterior probabilities of the variants)}
}
\value{
Predicted probability of covered
}
\description{
Build and predict corrected coverage using logistic GAM
}
\details{
This function builds a GAM model with log(claimed_coverage/(1-claimed_coverage)) (logit(claimed) as the predictor and the binary covered value as the response. This model is then used to accurately predict the coverage probability of the causal variant in the credible set
}
\author{
Anna Hutchinson
}
