% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{pvals_pp}
\alias{pvals_pp}
\title{pvals_pp}
\usage{
pvals_pp(pvals, f, type, N, s, W = 0.2)
}
\arguments{
\item{pvals}{p-values of SNPs}

\item{f}{Minor allele frequencies}

\item{type}{Type of experiment ('quant' or 'cc')}

\item{N}{Total sample size}

\item{s}{Proportion of cases (N1/N0+N1), ignored if type=='quant'}

\item{W}{Prior for the standard deviation of the effect size parameter, beta (W = 0.2 default)}
}
\value{
Posterior probabilities of null model (no genetic effect) and causality for each SNP
}
\description{
Posterior probabilities of causality from p-values
}
\details{
This function converts p-values to posterior probabilities of causality, including the null model of no genetic effects
}
\author{
Anna Hutchinson
}
