% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_cov_functions.R
\name{quick_corrcov_cs}
\alias{quick_corrcov_cs}
\title{quick_corrcov_cs}
\usage{
quick_corrcov_cs(thr, simulated.pps, pp)
}
\arguments{
\item{thr}{Threshold value to exceed}

\item{simulated.pps}{A list of matrices of simulated posterior probabilities (1 simulation per row) where each list element is for the corresponding SNP considered causal}

\item{pp}{The posterior probabilities of the original system}
}
\value{
A list of the credible set obtained using the specified threshold, the corrected coverage estimate of this credible set, the threshold the user specified and the size of the credible set (the sum of the pps of the variants)
}
\description{
quick_corrcov_cs
}
