% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{Var.data.cc}
\alias{Var.data.cc}
\title{Variance of the estimated effect size for case-control data}
\usage{
Var.data.cc(f, N, s)
}
\arguments{
\item{f}{Minor allele frequencies}

\item{N}{Total sample size (N0+N1)}

\item{s}{Proportion of cases (N1/N0+N1)}
}
\value{
Variance of estimated effect size \eqn{\hat{\beta}}, V.
}
\description{
Variance of the estimated effect size for case-control data
}
\author{
Claudia Giambartolomei
}
