% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_functions.R
\name{z0_pp}
\alias{z0_pp}
\title{Find PPs from Z-scores and MAFs}
\usage{
z0_pp(z, f, type, N, s, W = 0.2)
}
\arguments{
\item{z}{Marginal Z-scores of SNPs}

\item{f}{Minor allele frequencies}

\item{type}{Type of experiment ('quant' or 'cc')}

\item{N}{Total sample size}

\item{s}{Proportion of cases (N1/N0+N1), ignored if type=='quant'}

\item{W}{Prior for the standard deviation of the effect size parameter beta}
}
\value{
Posterior probabilities of null model (no genetic effect) and causality for each SNP
}
\description{
Posterior probabilities of causality from marginal Z-scores with prior SD as a parameter
}
\details{
This function converts Z-scores to posterior probabilities of causality, including the null model of no genetic effects
}
\examples{

set.seed(1)
nsnps = 100
N0 = 5000
N1 = 5000
z_scores <- rnorm(nsnps, 0, 3)

nhaps <- 1000
lag <- 5
maf.tmp <- runif(nsnps+lag, 0.05, 0.5) # common SNPs
laghaps <- do.call("cbind", lapply(maf.tmp, function(f) rbinom(nhaps,1,f)))
haps <- laghaps[,1:nsnps]
for(j in 1:lag)
   haps <- haps + laghaps[,(1:nsnps)+j]
haps <- round(haps/matrix(apply(haps,2,max),nhaps,nsnps,byrow=TRUE))
maf <- colMeans(haps)

z0_pp(z = z_scores, f = maf, type = "cc", N = N0+N1, s = N1/(N0+N1))

}
\author{
Anna Hutchinson
}
