% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_simulation_functions.R
\name{z_sim}
\alias{z_sim}
\title{z_sim}
\usage{
z_sim(Zj, Sigma, nrep)
}
\arguments{
\item{Zj}{Vector of joint z-scores (0s except at CV)}

\item{Sigma}{SNP correlation matrix}

\item{nrep}{Number of simulated z-scores}
}
\value{
Matrix of simulated z-scores, one simulation per row
}
\description{
Simulate marginal z-scores from the joint z-scores using \eqn{Z_m ~ MVN(E(Z_j),\Sigma)}
}
