% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_simulation_functions.R
\name{z_sim}
\alias{z_sim}
\title{z_sim}
\usage{
z_sim(Zj, Sigma, nrep)
}
\arguments{
\item{Zj}{Vector of joint Z-scores (a vector of 0s except at the causal variant)}

\item{Sigma}{SNP correlation matrix}

\item{nrep}{Number of Z-score systems to simulate}
}
\value{
Matrix of simulated z-scores, one simulation per row
}
\description{
Simulate marginal z-scores (\eqn{Z_m}) from the joint z-scores (\eqn{Z_j}) using \eqn{E(Z_m) = Z_j \times \Sigma} and
\eqn{Z* \sim MVN(E(Z_m), \Sigma)}
}
